Finished with handler test

+++ models/post.go
@@ -4,5 +4,5 @@ package models
 type Post struct {
 	ID      uint   `gorm:"primary_key" json:"id"`
 	Caption string `json:"caption" gorm:"not null"`
-	UserID  int64  `json:"user_id" gorm:"not null"`
+	UserID  uint   `json:"user_id" gorm:"not null"`
 }
